feat(alerts): Change rules/settings order in alerts page
Also extracts header into a shared component.

Closes SEN-435
