+++ src/sentry/static/sentry/app/views/settings/projectAlerts/projectAlertSettings.jsx
@@ -6,17 +6,12 @@ import {t} from 'app/locale';
 import Access from 'app/components/acl/access';
 import AlertLink from 'app/components/alertLink';
 import AsyncView from 'app/views/asyncView';
-import Button from 'app/components/button';
 import Form from 'app/views/settings/components/forms/form';
 import JsonForm from 'app/views/settings/components/forms/jsonForm';
-import ListLink from 'app/components/listLink';
-import NavTabs from 'app/components/navTabs';
 import PermissionAlert from 'app/views/settings/project/permissionAlert';
 import PluginList from 'app/components/pluginList';
 import SentryTypes from 'app/sentryTypes';
-import SettingsPageHeader from 'app/views/settings/components/settingsPageHeader';
-import Tooltip from 'app/components/tooltip';
-import recreateRoute from 'app/utils/recreateRoute';
+import ProjectAlertHeader from './projectAlertHeader';
 
 export default class ProjectAlertSettings extends AsyncView {
   static propTypes = {
@@ -71,43 +66,16 @@ export default class ProjectAlertSettings extends AsyncView {
   }
 
   renderBody() {
-    const {orgId, projectId} = this.props.params;
-    const {organization} = this.props;
-    const canEditRule = organization.access.includes('project:write');
+    const {
+      organization,
+      params: {orgId, projectId},
+    } = this.props;
 
     return (
       <Access access={['project:write']}>
         {({hasAccess}) => (
           <React.Fragment>
-            <SettingsPageHeader
-              title={t('Alerts')}
-              action={
-                <Tooltip
-                  disabled={canEditRule}
-                  title={t('You do not have permission to edit alert rules.')}
-                >
-                  <Button
-                    to={recreateRoute('rules/new/', this.props)}
-                    disabled={!canEditRule}
-                    priority="primary"
-                    size="small"
-                    icon="icon-circle-add"
-                  >
-                    {t('New Alert Rule')}
-                  </Button>
-                </Tooltip>
-              }
-              tabs={
-                <NavTabs underlined={true}>
-                  <ListLink to={recreateRoute('', this.props)} index={true}>
-                    {t('Settings')}
-                  </ListLink>
-                  <ListLink to={recreateRoute('rules/', this.props)}>
-                    {t('Rules')}
-                  </ListLink>
-                </NavTabs>
-              }
-            />
+            <ProjectAlertHeader projectId={projectId} />
             <PermissionAlert />
             <AlertLink to={'/settings/account/notifications/'} icon="icon-mail">
               {t(
